Add fixtures to tests

+++ ligotools/utils.py
@@ -11,7 +11,7 @@ from scipy.signal import butter, filtfilt, iirdesign, zpk2tf, freqz
 def whiten(strain, interp_psd, dt):
     Nt = len(strain)
     freqs = np.fft.rfftfreq(Nt, dt)
-    freqs1 = np.linspace(0,2048.,Nt/2+1)
+    freqs1 = np.linspace(0,2048., int(Nt/2)+1)
 
     # whitening: transform to freq domain, divide by asd, then transform back, 
     # taking care to get normalization right.
